feat: limit age to 18-65 year olds
Crazy how much this affects results. Children must have a large impact (and generally don't have much agency over own housing decisions). Will have to think through these implications carefully and make some individual scatter plots showing the correlation of each of these explanatory variables with the multifamily housing output variable, ideally with colored dots by race/ethnicity group and also with two plots - one with a relationship among children and one with adults (and perhaps with the 65+ group, to show that they are or are not responsible for the dramatic shifts caused by this filtering).

In short - lots to do!