fix: correct variables that need LINENO -> PERNUM translation
I previously implemented the correct logic but on the wrong columns. The CPS reports some family interrelationships with the target individual using the PERNUM column and some using the LINENO column. For example, if someone reports that person 5 in the household is their mother, does that refer to the person with PERNUM == 5 or LINENO == 5? Sometimes PERNUM and LINENO don't match so it's important to distinguish and translate between the two.

Eventually, I'll completely translate to PERNUM at the start and eliminate LINENO to avoid confusion.